Moved the action struct to a dto.
